requirements.txt
urllib3>=1.22
pillow>=6.2.1
argparse>=1.4.0
gpxpy>=1.3.5
img2pdf>=0.3.3
